pyproject.toml
[tool.poetry]
name = "reminders"
version = "0.1.0"
description = ""
authors = ["katewitkoski <ka.lwitkoski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
yoyo-migrations = "^8.2.0"
psycopg-binary = "^3.1.17"
python-dotenv = "^1.0.1"
psycopg = "^3.1.17"
uvicorn = "^0.27.0.post1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
